+++ src/screens/CartScreen/CartScreen.tsx
@@ -1,113 +1,91 @@
-import React, { useContext, useEffect, useState } from "react";
-import {
-  Text,
-  View,
-  StyleSheet,
-  KeyboardAvoidingView,
-  Image,
-  Platform,
-} from "react-native";
-import ButtonShop from "../../components/ButtonShop";
-import { StackScreenProps } from "@react-navigation/stack";
-import { PurchaseStackParamList } from "../../navigations/PurchaseNavigator";
-import ProductCartCard from "../../components/ProductCartCard";
-import InputNumber from "../../components/InputNumber";
-import { CartDataSource } from "../../datasource/CartDataSource";
-import {
-  ScrollView,
-  TouchableOpacity,
-  TouchableWithoutFeedback,
-} from "react-native-gesture-handler";
-import { ShopEntity } from "../../entities/ShopEntity";
-import Dash from "react-native-dash";
-import {
-  CartEntity,
-  ItemCart,
-  paymentCartEntity,
-  PremiumEntity,
-} from "../../entities/CartEntity";
-import { OrderFacade } from "../../facade/OrderFacade";
-import { OrderEntity } from "../../entities/OrderEntity";
-import { currencyFormat } from "../../utilities/CurrencyFormat";
-import SkeletonPlaceholder from "react-native-skeleton-placeholder";
-import CartEmptyState from "./CartEmptyState";
-import AccrodingPrice from "../../components/AccrodingPrice";
-import { useIsFocused } from "@react-navigation/native";
-import PremiumCard from "../../components/PremiumCard";
-import { AccrodionPriceModel } from "../../models/AccrodionPriceModel";
-import { CheckBox } from "native-base";
-import { UserDataContext } from "../../provider/UserDataProvider";
-import ShipmentSection from "./ShipmentSection";
-import { Shipment } from "./Shipment";
-import { CartContext } from "../../context/cartStore";
-import { Types } from "../../context/cartReducer";
-import PaymentSection from "./PaymentSection";
+import React, { useContext, useEffect, useState } from 'react'
+import { Text, View, StyleSheet, KeyboardAvoidingView, Image, Platform } from 'react-native'
+import ButtonShop from '../../components/ButtonShop'
+import { StackScreenProps } from '@react-navigation/stack'
+import { PurchaseStackParamList } from '../../navigations/PurchaseNavigator'
+import ProductCartCard from '../../components/ProductCartCard'
+import InputNumber from '../../components/InputNumber'
+import { CartDataSource } from '../../datasource/CartDataSource'
+import { FlatList, ScrollView, TouchableOpacity } from 'react-native-gesture-handler'
+import { ShopEntity } from '../../entities/ShopEntity'
+import Dash from 'react-native-dash'
+import { CartEntity, ItemCart, PremiumEntity } from '../../entities/CartEntity'
+import { OrderFacade } from '../../facade/OrderFacade'
+import { OrderEntity } from '../../entities/OrderEntity'
+import { currencyFormat } from '../../utilities/CurrencyFormat'
+import SkeletonPlaceholder from 'react-native-skeleton-placeholder'
+import CartEmptyState from './CartEmptyState'
+import AccrodingPrice from '../../components/AccrodingPrice'
+import { useIsFocused } from '@react-navigation/native'
+import PremiumCard from '../../components/PremiumCard'
+import { AccrodionPriceModel } from '../../models/AccrodionPriceModel'
+import { CheckBox } from 'native-base'
+import { UserDataContext } from '../../provider/UserDataProvider'
+import ShipmentSection from './ShipmentSection'
+import { Shipment } from './Shipment'
+import { CartContext } from '../../context/cartStore'
+import { Types } from '../../context/cartReducer'
+import PaymentSection from './PaymentSection'
 
-type ShopScreenRouteProp = StackScreenProps<PurchaseStackParamList, "Cart">;
+type ShopScreenRouteProp = StackScreenProps<PurchaseStackParamList, 'Cart'>
 
 const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
-  const [shipment, setShipment] = useState<Shipment>();
-  const [cart, setCart] = useState<CartEntity | undefined>();
-  const [quantity, setQuantity] = useState(0);
-  const [payment, setPayment] = useState<string | undefined>();
-  const [specialRequest, setSpecialRequest] = useState<AccrodionPriceModel[]>(
-    []
-  );
-  const [discoutPromo, setDiscoutPromo] = useState<AccrodionPriceModel[]>([]);
-  const [useSubsidize, setUseSubsudize] = useState(false);
-  const isFocused = useIsFocused();
-  const userDataStore = useContext(UserDataContext);
-  const { userData } = userDataStore;
-  const { dispatch } = useContext(CartContext);
+  const [shipment, setShipment] = useState<Shipment>()
+  const [cart, setCart] = useState<CartEntity | undefined>()
+  const [quantity, setQuantity] = useState(0)
+  const [payment, setPayment] = useState<string | undefined>()
+  const [specialRequest, setSpecialRequest] = useState<AccrodionPriceModel[]>([])
+  const [discoutPromo, setDiscoutPromo] = useState<AccrodionPriceModel[]>([])
+  const [useSubsidize, setUseSubsudize] = useState(false)
+  const isFocused = useIsFocused()
+  const userDataStore = useContext(UserDataContext)
+  const { userData } = userDataStore
+  const { dispatch } = useContext(CartContext)
 
   useEffect(() => {
-    getCart();
-  }, [isFocused]);
+    getCart()
+  }, [isFocused])
 
   // TODO: don't use any type
   const formatAccrodion = (data: any[]): AccrodionPriceModel[] => {
-    let arrayOutput: any[] = [];
+    let arrayOutput: any[] = []
     data.map((item: any) => {
-      let unit = item.sale_unit ? item.sale_unit : item.unit;
+      let unit = item.sale_unit ? item.sale_unit : item.unit
       arrayOutput.push({
         item: `${item.name} (${item.price}฿ x ${item.quantity} ${unit})`,
         price: item.price,
         quantity: item.quantity,
-      });
-    });
-    return arrayOutput;
-  };
+      })
+    })
+    return arrayOutput
+  }
 
   const getCart = async () => {
-    CartDataSource.getCartByShop(
-      route.params.company,
-      route.params.shop.id,
-      route.params.productBrand
-    ).then((res: CartEntity) => {
-      setCart(res);
-      let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
-        res.received_special_request_discounts
-      );
-      let discountProduct: AccrodionPriceModel[] = formatAccrodion(
-        res.received_discounts.filter((item) => item.item_id != null)
-      );
-      setSpecialRequest(discountSpecial);
-      setDiscoutPromo(discountProduct);
-      setUseSubsudize(false);
-      if (res?.available_payments && res?.available_payments.length <= 1) {
-        res?.available_payments.map((item) => {
-          setPayment(item.id);
-          CartDataSource.calculate(
-            route.params.company,
-            route.params.shop.id,
-            item.id,
-            useSubsidize,
-            route.params.productBrand
-          ).then((res) => setCart(res));
-        });
-      }
-    });
-  };
+    CartDataSource.getCartByShop(route.params.company, route.params.shop.id, route.params.productBrand).then(
+      (res: CartEntity) => {
+        setCart(res)
+        let discountSpecial: AccrodionPriceModel[] = formatAccrodion(res.received_special_request_discounts)
+        let discountProduct: AccrodionPriceModel[] = formatAccrodion(
+          res.received_discounts.filter((item) => item.item_id != null),
+        )
+        setSpecialRequest(discountSpecial)
+        setDiscoutPromo(discountProduct)
+        setUseSubsudize(false)
+        if (res?.available_payments && res?.available_payments.length <= 1) {
+          res?.available_payments.map((item) => {
+            setPayment(item.id)
+            CartDataSource.calculate(
+              route.params.company,
+              route.params.shop.id,
+              item.id,
+              useSubsidize,
+              route.params.productBrand,
+            ).then((res) => setCart(res))
+          })
+        }
+      },
+    )
+  }
 
   const increaseProduct = async (itemId: string, quantity: number) => {
     dispatch({
@@ -117,7 +95,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         quantity: quantity + 1,
         shopId: route.params.shop.id,
       },
-    });
+    })
     await CartDataSource.addToCartByShopId(
       route.params.company,
       route.params.shop.id,
@@ -125,20 +103,18 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
       quantity + 1,
       payment,
       useSubsidize,
-      route.params.productBrand
+      route.params.productBrand,
     ).then((res: CartEntity) => {
-      setCart(res);
+      setCart(res)
 
-      let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
-        res.received_special_request_discounts
-      );
+      let discountSpecial: AccrodionPriceModel[] = formatAccrodion(res.received_special_request_discounts)
       let discountProduct: AccrodionPriceModel[] = formatAccrodion(
-        res.received_discounts.filter((item) => item.item_id != null)
-      );
-      setSpecialRequest(discountSpecial);
-      setDiscoutPromo(discountProduct);
-    });
-  };
+        res.received_discounts.filter((item) => item.item_id != null),
+      )
+      setSpecialRequest(discountSpecial)
+      setDiscoutPromo(discountProduct)
+    })
+  }
 
   const decreaseProduct = async (itemId: string, quantity: number) => {
     await CartDataSource.addToCartByShopId(
@@ -148,23 +124,21 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
       quantity - 1,
       payment,
       useSubsidize,
-      route.params.productBrand
+      route.params.productBrand,
     ).then((res: CartEntity) => {
-      setCart(res);
-      let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
-        res.received_special_request_discounts
-      );
+      setCart(res)
+      let discountSpecial: AccrodionPriceModel[] = formatAccrodion(res.received_special_request_discounts)
       let discountProduct: AccrodionPriceModel[] = formatAccrodion(
-        res.received_discounts.filter((item) => item.item_id != null)
-      );
-      setSpecialRequest(discountSpecial);
-      setDiscoutPromo(discountProduct);
-    });
-  };
+        res.received_discounts.filter((item) => item.item_id != null),
+      )
+      setSpecialRequest(discountSpecial)
+      setDiscoutPromo(discountProduct)
+    })
+  }
 
   const adjustProduct = async (itemId: string, quantity: number) => {
-    const regexp = /^[0-9\b]+$/;
-    if (quantity.toString() === "" || regexp.test(quantity.toString())) {
+    const regexp = /^[0-9\b]+$/
+    if (quantity.toString() === '' || regexp.test(quantity.toString())) {
       CartDataSource.addToCartByShopId(
         route.params.company,
         route.params.shop.id,
@@ -172,22 +146,20 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         quantity,
         payment,
         useSubsidize,
-        route.params.productBrand
+        route.params.productBrand,
       ).then((res: CartEntity) => {
-        setCart(res);
-        let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
-          res.received_special_request_discounts
-        );
+        setCart(res)
+        let discountSpecial: AccrodionPriceModel[] = formatAccrodion(res.received_special_request_discounts)
         let discountProduct: AccrodionPriceModel[] = formatAccrodion(
-          res.received_discounts.filter((item) => item.item_id != null)
-        );
-        setSpecialRequest(discountSpecial);
-        setDiscoutPromo(discountProduct);
-      });
+          res.received_discounts.filter((item) => item.item_id != null),
+        )
+        setSpecialRequest(discountSpecial)
+        setDiscoutPromo(discountProduct)
+      })
     } else {
-      alert("Number Only");
+      alert('Number Only')
     }
-  };
+  }
 
   const removeItem = async (itemId: string) => {
     dispatch({
@@ -196,91 +168,81 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         id: itemId,
         shopId: route.params.shop.id,
       },
-    });
+    })
     CartDataSource.removeItem(
       route.params.company,
       route.params.shop.id,
       itemId,
       payment,
       useSubsidize,
-      route.params.productBrand
+      route.params.productBrand,
     ).then((res: CartEntity) => {
-      setCart(res);
-      let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
-        res.received_special_request_discounts
-      );
+      setCart(res)
+      let discountSpecial: AccrodionPriceModel[] = formatAccrodion(res.received_special_request_discounts)
       let discountProduct: AccrodionPriceModel[] = formatAccrodion(
-        res.received_discounts.filter((item) => item.item_id != null)
-      );
-      setSpecialRequest(discountSpecial);
-      setDiscoutPromo(discountProduct);
-    });
-  };
+        res.received_discounts.filter((item) => item.item_id != null),
+      )
+      setSpecialRequest(discountSpecial)
+      setDiscoutPromo(discountProduct)
+    })
+  }
 
   const handleShipmentChange = (s: Shipment) => {
-    setShipment(s);
-  };
+    setShipment(s)
+  }
 
   const handlePayment = (payment: string) => {
-    if (payment == "cash") {
-      setPayment("cash");
+    if (payment == 'cash') {
+      setPayment('cash')
       CartDataSource.calculate(
         route.params.company,
         route.params.shop.id,
         payment,
         useSubsidize,
-        route.params.productBrand
+        route.params.productBrand,
       ).then((res: CartEntity) => {
-        setCart(res);
-        let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
-          res.received_special_request_discounts
-        );
+        setCart(res)
+        let discountSpecial: AccrodionPriceModel[] = formatAccrodion(res.received_special_request_discounts)
         let discountProduct: AccrodionPriceModel[] = formatAccrodion(
-          res.received_discounts.filter((item) => item.item_id != null)
-        );
-        setSpecialRequest(discountSpecial);
-        setDiscoutPromo(discountProduct);
-      });
+          res.received_discounts.filter((item) => item.item_id != null),
+        )
+        setSpecialRequest(discountSpecial)
+        setDiscoutPromo(discountProduct)
+      })
     } else {
-      setPayment("credit");
+      setPayment('credit')
       CartDataSource.calculate(
         route.params.company,
         route.params.shop.id,
         payment,
         useSubsidize,
-        route.params.productBrand
+        route.params.productBrand,
       ).then((res: CartEntity) => {
-        setCart(res);
-        let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
-          res.received_special_request_discounts
-        );
+        setCart(res)
+        let discountSpecial: AccrodionPriceModel[] = formatAccrodion(res.received_special_request_discounts)
         let discountProduct: AccrodionPriceModel[] = formatAccrodion(
-          res.received_discounts.filter((item) => item.item_id != null)
-        );
-        setSpecialRequest(discountSpecial);
-        setDiscoutPromo(discountProduct);
-      });
+          res.received_discounts.filter((item) => item.item_id != null),
+        )
+        setSpecialRequest(discountSpecial)
+        setDiscoutPromo(discountProduct)
+      })
     }
-  };
+  }
 
   const handleUseSubsidize = (b: boolean) => {
-    CartDataSource.calculate(
-      route.params.company,
-      route.params.shop.id,
-      payment,
-      b,
-      route.params.productBrand
-    ).then((res: CartEntity) => {
-      setCart(res);
-      setUseSubsudize(b);
-    });
-  };
+    CartDataSource.calculate(route.params.company, route.params.shop.id, payment, b, route.params.productBrand).then(
+      (res: CartEntity) => {
+        setCart(res)
+        setUseSubsudize(b)
+      },
+    )
+  }
 
   const confirmOrder = (shop: ShopEntity, cart?: CartEntity) => {
     if (!cart?.selected_payment) {
-      alert("กรุณาเลือกวิธีการชำระเงิน");
+      alert('กรุณาเลือกวิธีการชำระเงิน')
     } else if (!shipment) {
-      alert("กรุณาเลือกสถานที่จัดส่ง");
+      alert('กรุณาเลือกสถานที่จัดส่ง')
     } else {
       OrderFacade.confirmOrder(
         userData.company,
@@ -288,40 +250,32 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         shipment,
         cart,
         cart.subsidize_discount,
-        route.params.productBrand
+        route.params.productBrand,
       ).then((res: OrderEntity) => {
-        CartDataSource.clearCart(route.params.company, shop.id, route.params.productBrand);
-        navigation.navigate("OrderSuccess", { data: res, cart });
-      });
+        CartDataSource.clearCart(route.params.company, shop.id, route.params.productBrand)
+        navigation.navigate('OrderSuccess', { data: res, cart })
+      })
     }
-  };
+  }
 
   const getCashDiscount = (cart: CartEntity): number => {
-    let discountItem = cart.received_discounts.find((i) => i.id === "cash");
-    return discountItem ? discountItem.price * discountItem.quantity : 0;
-  };
+    let discountItem = cart.received_discounts.find((i) => i.id === 'cash')
+    return discountItem ? discountItem.price * discountItem.quantity : 0
+  }
 
-  const getPromoDiscountForItem = (
-    cart: CartEntity,
-    itemId: string
-  ): number => {
-    let discountItem = cart.received_discounts.find(
-      (i) => i.item_id === itemId
-    );
-    return discountItem ? discountItem.price * discountItem.quantity : 0;
-  };
+  const getPromoDiscountForItem = (cart: CartEntity, itemId: string): number => {
+    let discountItem = cart.received_discounts.find((i) => i.item_id === itemId)
+    return discountItem ? discountItem.price * discountItem.quantity : 0
+  }
 
   return (
     <KeyboardAvoidingView
       style={styled.container}
-      behavior={Platform.OS == "ios" ? "padding" : "height"}
+      behavior={Platform.OS == 'ios' ? 'padding' : 'height'}
       keyboardVerticalOffset={80}
     >
       <View style={styled.warpChangeShop}>
-        <ButtonShop
-          shopName={route.params.shop.name}
-          onPress={() => navigation.navigate("ShopList")}
-        />
+        <ButtonShop shopName={route.params.shop.name} onPress={() => navigation.navigate('ShopList')} />
       </View>
       {cart !== undefined ? (
         <View style={styled.borderContainer}>
@@ -331,7 +285,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                 <View style={styled.productContainer}>
                   <Text style={styled.textProduct}>สินค้า</Text>
                   {cart.items.map((item: ItemCart, index: number) => {
-                    let discount = getPromoDiscountForItem(cart, item.id);
+                    let discount = getPromoDiscountForItem(cart, item.id)
                     return (
                       <ProductCartCard
                         key={item.id}
@@ -353,50 +307,44 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                           key={item.title}
                           value={item.quantity.toString()}
                           onPlus={() => increaseProduct(item.id, item.quantity)}
-                          onMinus={() =>
-                            decreaseProduct(item.id, item.quantity)
-                          }
+                          onMinus={() => decreaseProduct(item.id, item.quantity)}
                           onChangeText={(e: any) => {
-                            setQuantity((cart.items[index].quantity = e));
+                            setQuantity((cart.items[index].quantity = e))
                           }}
                           onBlur={() => adjustProduct(item.id, quantity)}
                         />
                       </ProductCartCard>
-                    );
+                    )
                   })}
                 </View>
-                {cart.available_premiums.length > 0 ? (
-                  <View style={styled.PremiumContainer}>
-                    <Text style={{ fontSize: 18, fontWeight: "bold" }}>
-                      ของแถมที่ได้รับ
-                    </Text>
-                    {cart.available_premiums.map((item: PremiumEntity) => {
-                      return (
-                        <PremiumCard
-                          title={item.name}
-                          desc={item.packing_size}
-                          image={item.image}
-                          quantity={item.quantity}
-                          unit={item.unit}
-                        />
-                      );
-                    })}
-                  </View>
-                ) : null}
+                <FlatList
+                  horizontal
+                  showsHorizontalScrollIndicator={false}
+                  data={cart.available_premiums}
+                  style={{ backgroundColor: "#FFFFFF" ,marginHorizontal:4}}
+                  renderItem={({ item }) => (
+                    <PremiumCard
+                      title={item.name}
+                      desc={item.packing_size}
+                      image={item.image}
+                      quantity={item.quantity}
+                      unit={item.unit}
+                    />
+                  )}
+                  keyExtractor={(item) => item.id}
+                />
                 {specialRequest.length > 0 ? (
                   <View style={styled.specialRequestContainer}>
                     <View
                       style={{
-                        flexDirection: "row",
-                        justifyContent: "space-between",
+                        flexDirection: 'row',
+                        justifyContent: 'space-between',
                       }}
                     >
-                      <Text style={styled.textHeaderPayment}>
-                        Special Request
-                      </Text>
+                      <Text style={styled.textHeaderPayment}>Special Request</Text>
                       <TouchableOpacity
                         onPress={() =>
-                          navigation.navigate("SpecialRequest", {
+                          navigation.navigate('SpecialRequest', {
                             cart: cart,
                             shop: route.params.shop,
                             item: specialRequest,
@@ -407,8 +355,8 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                       >
                         <Text
                           style={{
-                            color: "#4C95FF",
-                            fontWeight: "500",
+                            color: '#4C95FF',
+                            fontWeight: '500',
                             fontSize: 14,
                           }}
                         >
@@ -421,13 +369,13 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                       title="ขอส่วนลดพิเศษเพิ่ม"
                       total={cart.total_received_special_request_discount}
                       detail={specialRequest}
-                      price_color={"#BB6BD9"}
+                      price_color={'#BB6BD9'}
                     />
                   </View>
                 ) : (
                   <TouchableOpacity
                     onPress={() =>
-                      navigation.navigate("SpecialRequest", {
+                      navigation.navigate('SpecialRequest', {
                         cart: cart,
                         shop: route.params.shop,
                         item: specialRequest,
@@ -440,11 +388,9 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                     <View style={styled.buttonSpecialRequest}>
                       <Image
                         style={{ width: 25, height: 25 }}
-                        source={require("../../../assets/special_request.png")}
+                        source={require('../../../assets/special_request.png')}
                       />
-                      <Text style={styled.textButtonSpecialRequest}>
-                        Special Request
-                      </Text>
+                      <Text style={styled.textButtonSpecialRequest}>Special Request</Text>
                     </View>
                   </TouchableOpacity>
                 )}
@@ -453,54 +399,47 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                   <View
                     style={{
                       paddingLeft: 20,
-                      backgroundColor: "#FFFFFF",
+                      backgroundColor: '#FFFFFF',
                     }}
                   >
                     <Text style={styled.textHeaderPayment}>หมายเหตุ</Text>
                     <View
                       style={{
-                        flexDirection: "row",
+                        flexDirection: 'row',
                         marginVertical: 12,
-                        alignItems: "center",
+                        alignItems: 'center',
                       }}
                     >
-                      <Text style={{ color: "#6B7995" }}>
-                        {cart.special_request_remark}
-                      </Text>
+                      <Text style={{ color: '#6B7995' }}>{cart.special_request_remark}</Text>
                     </View>
                   </View>
                 ) : null}
                 <PaymentSection
                   payments={cart.available_payments}
                   onChange={(e) => {
-                    handlePayment(e);
+                    handlePayment(e)
                   }}
                 />
-                <View
-                  style={{ paddingHorizontal: 15, backgroundColor: "white" }}
-                >
+                <View style={{ paddingHorizontal: 15, backgroundColor: 'white' }}>
                   <View
                     style={{
-                      flexDirection: "row",
-                      justifyContent: "space-between",
-                      alignItems: "center",
+                      flexDirection: 'row',
+                      justifyContent: 'space-between',
+                      alignItems: 'center',
                     }}
                   >
                     <Text style={styled.textHeaderPayment}>ส่วนลดดูแลราคา</Text>
-                    <Text style={{ color: "#616A7B" }}>
+                    <Text style={{ color: '#616A7B' }}>
                       {useSubsidize
-                        ? currencyFormat(
-                            cart.available_subsidize - cart.usable_subsidize,
-                            2
-                          )
+                        ? currencyFormat(cart.available_subsidize - cart.usable_subsidize, 2)
                         : currencyFormat(cart.available_subsidize, 2)}
                     </Text>
                   </View>
                   <View
                     style={{
-                      flexDirection: "row",
+                      flexDirection: 'row',
                       marginVertical: 12,
-                      alignItems: "center",
+                      alignItems: 'center',
                     }}
                   >
                     {cart.available_subsidize > 0 ? (
@@ -513,20 +452,16 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                             borderRadius: 4,
                             width: 20,
                             height: 20,
-                            alignItems: "center",
+                            alignItems: 'center',
                           }}
                         />
-                        <Text style={{ marginLeft: 15, color: "#6B7995" }}>
-                          ใช้ส่วนลด
-                        </Text>
-                        <Text style={{ color: "#FF5D5D", fontWeight: "bold" }}>
-                          {" " + currencyFormat(cart.usable_subsidize, 2)}
+                        <Text style={{ marginLeft: 15, color: '#6B7995' }}>ใช้ส่วนลด</Text>
+                        <Text style={{ color: '#FF5D5D', fontWeight: 'bold' }}>
+                          {' ' + currencyFormat(cart.usable_subsidize, 2)}
                         </Text>
                       </>
                     ) : (
-                      <Text style={{ color: "#6B7995" }}>
-                        ไม่มีวงเงินที่สามารถใช้ได้
-                      </Text>
+                      <Text style={{ color: '#6B7995' }}>ไม่มีวงเงินที่สามารถใช้ได้</Text>
                     )}
                   </View>
                 </View>
@@ -534,27 +469,23 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                   dashGap={2}
                   dashLength={4}
                   dashThickness={4}
-                  style={{ width: "100%", height: 1 }}
+                  style={{ width: '100%', height: 1 }}
                   dashColor="#C8CDD6"
                 />
                 <View style={styled.totalPriceContainer}>
                   <View style={styled.warpPrice}>
                     <Text style={styled.textDiscount}>ราคาก่อนลด</Text>
-                    <Text style={styled.textBeforeDiscount}>
-                      {currencyFormat(cart.before_discount, 2)}
-                    </Text>
+                    <Text style={styled.textBeforeDiscount}>{currencyFormat(cart.before_discount, 2)}</Text>
                   </View>
 
-                  {cart.received_discounts.filter(
-                    (item) => item.item_id != null
-                  ).length > 0 ? (
+                  {cart.received_discounts.filter((item) => item.item_id != null).length > 0 ? (
                     <AccrodingPrice
                       title="ส่วนลดรายการ"
                       total={cart?.received_discounts
                         .filter((item) => item.item_id != null)
                         .reduce((sum, item) => sum + item.total, 0)}
                       detail={discoutPromo}
-                      price_color={"#3AAE49"}
+                      price_color={'#3AAE49'}
                     />
                   ) : null}
                   {cart.received_special_request_discounts.length > 0 ? (
@@ -562,15 +493,13 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                       title="ขอส่วนลดพิเศษเพิ่ม"
                       total={cart.total_received_special_request_discount}
                       detail={specialRequest}
-                      price_color={"#BB6BD9"}
+                      price_color={'#BB6BD9'}
                     />
                   ) : null}
                   {cart.subsidize_discount != 0 ? (
                     <View style={styled.warpPrice}>
                       <Text style={styled.textDiscount}>ส่วนลดดูแลราคา</Text>
-                      <Text style={styled.textSubsidizeDiscount}>
-                        {currencyFormat(cart.subsidize_discount, 2)}
-                      </Text>
+                      <Text style={styled.textSubsidizeDiscount}>{currencyFormat(cart.subsidize_discount, 2)}</Text>
                     </View>
                   ) : null}
                   {getCashDiscount(cart) != 0 ? (
@@ -578,9 +507,9 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                       <Text style={styled.textDiscount}>ส่วนลดเงินสด</Text>
                       <Text
                         style={{
-                          color: "#FF8329",
+                          color: '#FF8329',
                           fontSize: 16,
-                          fontWeight: "bold",
+                          fontWeight: 'bold',
                         }}
                       >
                         {currencyFormat(getCashDiscount(cart), 2)}
@@ -589,23 +518,19 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                   ) : null}
                   <View
                     style={{
-                      backgroundColor: "#FBFBFB",
-                      flexDirection: "row",
-                      justifyContent: "space-between",
+                      backgroundColor: '#FBFBFB',
+                      flexDirection: 'row',
+                      justifyContent: 'space-between',
                       marginVertical: 5,
                       padding: 5,
                     }}
                   >
                     <Text style={styled.textBeforeTotal}>ส่วนลดรวม</Text>
-                    <Text style={styled.textTotalDiscount}>
-                      {currencyFormat(cart.total_discount, 2)}
-                    </Text>
+                    <Text style={styled.textTotalDiscount}>{currencyFormat(cart.total_discount, 2)}</Text>
                   </View>
                   <View style={styled.warpPrice}>
                     <Text style={styled.textLabelTotalPrice}>ราคารวม</Text>
-                    <Text style={styled.textTotalPrice}>
-                      {currencyFormat(cart.total_price, 2)}
-                    </Text>
+                    <Text style={styled.textTotalPrice}>{currencyFormat(cart.total_price, 2)}</Text>
                   </View>
                 </View>
                 <View style={styled.warpDelivery}>
@@ -613,7 +538,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                     company={route.params.company}
                     shopId={route.params.shop.id}
                     onChange={(v) => {
-                      if (v) handleShipmentChange(v);
+                      if (v) handleShipmentChange(v)
                     }}
                     withDefault={true}
                   />
@@ -621,11 +546,11 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
               </ScrollView>
               <View
                 style={{
-                  justifyContent: "center",
-                  backgroundColor: "#FFFFFF",
-                  flexDirection: "column",
+                  justifyContent: 'center',
+                  backgroundColor: '#FFFFFF',
+                  flexDirection: 'column',
                   height: 90,
-                  shadowColor: "#000",
+                  shadowColor: '#000',
                   shadowOffset: {
                     width: 0,
                     height: 6,
@@ -638,18 +563,13 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                 <TouchableOpacity
                   style={styled.confirmOrderButton}
                   onPress={() => {
-                    confirmOrder(route.params.shop, cart);
+                    confirmOrder(route.params.shop, cart)
                   }}
                 >
                   <View style={styled.iconCartWarp}>
-                    <Image
-                      style={styled.iconLocation}
-                      source={require("../../../assets/order-cart.png")}
-                    />
+                    <Image style={styled.iconLocation} source={require('../../../assets/order-cart.png')} />
                   </View>
-                  <Text style={styled.textconfirmOrderButton}>
-                    ยืนยันคำสั่งซื้อ
-                  </Text>
+                  <Text style={styled.textconfirmOrderButton}>ยืนยันคำสั่งซื้อ</Text>
                 </TouchableOpacity>
               </View>
             </>
@@ -662,7 +582,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
           <View style={styled.productContainer}>
             <Text style={styled.textProduct}>สินค้า</Text>
             <SkeletonPlaceholder>
-              <View style={{ flexDirection: "row", alignItems: "center" }}>
+              <View style={{ flexDirection: 'row', alignItems: 'center' }}>
                 <View style={{ width: 60, height: 60 }} />
                 <View style={{ marginLeft: 20 }}>
                   <View style={{ width: 120, height: 20, borderRadius: 4 }} />
@@ -681,22 +601,22 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         </>
       )}
     </KeyboardAvoidingView>
-  );
-};
+  )
+}
 
-export default CartScreen;
+export default CartScreen
 
 const styled = StyleSheet.create({
-  container: { flex: 1, backgroundColor: "#F8FAFF" },
+  container: { flex: 1, backgroundColor: '#F8FAFF' },
   warpChangeShop: {
     padding: 10,
   },
   borderContainer: {
     flex: 2,
-    backgroundColor: "#F8FAFF",
+    backgroundColor: '#F8FAFF',
   },
   productContainer: {
-    backgroundColor: "white",
+    backgroundColor: 'white',
     padding: 15,
   },
   innerProductContainer: { marginTop: 10 },
@@ -704,134 +624,134 @@ const styled = StyleSheet.create({
   paymentContainer: {
     marginTop: 10,
     padding: 15,
-    backgroundColor: "white",
+    backgroundColor: 'white',
   },
   paymentMethod: {
     marginTop: 10,
-    borderBottomColor: "#EBEFF2",
+    borderBottomColor: '#EBEFF2',
     borderBottomWidth: 2,
-    borderTopColor: "#EBEFF2",
+    borderTopColor: '#EBEFF2',
     borderTopWidth: 2,
     borderRadius: 3,
     padding: 20,
   },
   textHeaderPayment: {
     fontSize: 18,
-    fontWeight: "bold",
+    fontWeight: 'bold',
   },
   textBodyPayment: {
     fontSize: 16,
-    color: "#616A7B",
+    color: '#616A7B',
   },
   lineDash: {
-    borderStyle: "dashed",
+    borderStyle: 'dashed',
     borderWidth: 5,
     borderRadius: 5,
-    borderColor: "#C8CDD6",
+    borderColor: '#C8CDD6',
     borderTopWidth: 5,
     borderTopLeftRadius: 5,
     borderTopRightRadius: 5,
   },
   line: {
     marginTop: 10,
-    borderBottomColor: "#EBEFF2",
+    borderBottomColor: '#EBEFF2',
     borderBottomWidth: 2,
     borderRadius: 3,
   },
   totalPriceContainer: {
     padding: 15,
-    backgroundColor: "white",
+    backgroundColor: 'white',
   },
   warpPrice: {
-    flexDirection: "row",
-    justifyContent: "space-between",
+    flexDirection: 'row',
+    justifyContent: 'space-between',
     marginVertical: 5,
   },
   textBeforeDiscount: {
-    color: "#6B7995",
+    color: '#6B7995',
     fontSize: 16,
   },
   textSubsidizeDiscount: {
-    color: "#FF5D5D",
+    color: '#FF5D5D',
     fontSize: 16,
-    fontWeight: "bold",
+    fontWeight: 'bold',
   },
   textDiscountFromProduct: {
-    color: "rgba(58, 174, 73, 1)",
+    color: 'rgba(58, 174, 73, 1)',
     fontSize: 14,
-    fontWeight: "bold",
+    fontWeight: 'bold',
   },
   textDiscountFromCash: {
-    color: "#4C95FF",
+    color: '#4C95FF',
     fontSize: 16,
-    fontWeight: "bold",
+    fontWeight: 'bold',
   },
   textTotalDiscount: {
-    color: "#616A7B",
+    color: '#616A7B',
     fontSize: 16,
-    fontWeight: "bold",
+    fontWeight: 'bold',
   },
-  textDiscount: { fontSize: 14, color: "#6B7995" },
-  textBeforeTotal: { fontSize: 16, color: "#6B7995" },
-  textLabelTotalPrice: { fontSize: 16, color: "#314364", fontWeight: "bold" },
-  textTotalPrice: { fontSize: 20, color: "#4C95FF", fontWeight: "bold" },
+  textDiscount: { fontSize: 14, color: '#6B7995' },
+  textBeforeTotal: { fontSize: 16, color: '#6B7995' },
+  textLabelTotalPrice: { fontSize: 16, color: '#314364', fontWeight: 'bold' },
+  textTotalPrice: { fontSize: 20, color: '#4C95FF', fontWeight: 'bold' },
   confirmOrderButton: {
-    flexDirection: "row",
+    flexDirection: 'row',
     borderRadius: 6,
-    backgroundColor: "#4C95FF",
+    backgroundColor: '#4C95FF',
     padding: 10,
     margin: 20,
-    alignItems: "center",
+    alignItems: 'center',
     height: 50,
   },
-  iconLocation: { width: 20, height: 20, resizeMode: "contain" },
+  iconLocation: { width: 20, height: 20, resizeMode: 'contain' },
   iconCartWarp: {
     marginRight: 100,
     marginLeft: 10,
   },
   textconfirmOrderButton: {
-    color: "#FFFFFF",
+    color: '#FFFFFF',
     fontSize: 18,
-    fontWeight: "bold",
+    fontWeight: 'bold',
   },
   buttonSpecialRequestContainer: {
-    backgroundColor: "#FFFFFF",
+    backgroundColor: '#FFFFFF',
     marginTop: 10,
     padding: 20,
   },
   buttonSpecialRequest: {
-    backgroundColor: "#EAF4FF",
+    backgroundColor: '#EAF4FF',
     height: 48,
     borderRadius: 6,
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
   },
   textButtonSpecialRequest: {
     fontSize: 16,
-    color: "#4C95FF",
-    fontWeight: "bold",
+    color: '#4C95FF',
+    fontWeight: 'bold',
     marginLeft: 10,
   },
   PremiumContainer: {
-    backgroundColor: "#FFFFFF",
+    backgroundColor: '#FFFFFF',
     marginTop: 10,
     padding: 20,
   },
   specialRequestContainer: {
     marginTop: 10,
     padding: 20,
-    backgroundColor: "#FFFFFF",
+    backgroundColor: '#FFFFFF',
   },
   warpDelivery: {
     marginTop: 10,
-    backgroundColor: "#FFFFFF",
+    backgroundColor: '#FFFFFF',
     marginBottom: 10,
   },
   iconPin: {
     borderRadius: 50,
     borderWidth: 5,
-    borderColor: "#4C95FF",
+    borderColor: '#4C95FF',
     width: 20,
     height: 20,
     marginRight: 5,
@@ -839,13 +759,13 @@ const styled = StyleSheet.create({
   iconUnPin: {
     borderRadius: 50,
     borderWidth: 1,
-    borderColor: "#E5E5E5",
+    borderColor: '#E5E5E5',
     width: 20,
     height: 20,
     marginRight: 5,
   },
   methodChoiceContainer: {
-    flexDirection: "row",
+    flexDirection: 'row',
     marginBottom: 5,
   },
-});
+})